requirements.txt
boto3>=1.34.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.7.0
rich>=13.7.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.24.0
